Moving our Jinja2 filter to ``flask_assets``.
And having it override the build-in ``jinja2`` template, so that we get the
included Flask contexts by default.
